+++ src/lang/en.ts
@@ -3,7 +3,6 @@ import {Lang} from "./Lang";
 export const en: Lang = {
 	code: 'en',
 	display: 'English',
-	flag: '🇬🇧',
 	numbers: [
 		'zero',
 		'one',